requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
plotly==5.18.0
pdfplumber==0.10.3
pytesseract==0.3.10
opencv-python==4.9.0.80
mne==1.6.1
requests==2.31.0
beautifulsoup4==4.12.3
joblib==1.3.2
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
